newsCrawlerService/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autocuda==0.8
beautifulsoup4==4.9.3
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
cssselect==1.1.0
cymem==2.0.5
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
findfile==1.4rc0
Flask==2.0.1
gitdb==4.0.7
GitPython==3.1.18
google==3.0.0
google-api-core==1.31.0
google-auth==1.33.1
google-cloud-core==1.7.1
google-cloud-language==2.2.2
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
googledrivedownloader==0.4
grpcio==1.39.0
huggingface-hub==0.0.12
idna==3.2
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.1
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
murmurhash==1.0.5
networkx==2.5.1
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pathy==0.6.0
Pillow==8.3.1
preshed==3.0.5
proto-plus==1.19.0
protobuf==3.17.3
pyabsa==1.1.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.7.6
requests==2.26.0
requests-file==1.5.1
rq==1.9.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
Senta==1.0.3
seqeval==1.2.2
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.1.0
smmap==4.0.0
soupsieve==2.2.1
spacy==3.1.1
spacy-legacy==3.0.8
srsly==2.4.1
termcolor==1.1.0
thinc==8.0.8
threadpoolctl==2.2.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.10.3
torch==1.9.0
tqdm==4.61.2
transformers==4.9.0
typer==0.3.2
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
waitress==2.0.0
wasabi==0.8.2
Werkzeug==2.0.1
yarl==1.6.3
